Cooperation and cheating among Dictyostelium amoebae

This data describes the sporulation success of Dictyostelium discoideum amoebae in mixed-genotype fruiting bodies.

Usage

data_Madgwick_2018

Format

data_Madgwick_2018 is a data frame with 1226 rows and 7 columns:

strain_i

Name of strain i

strain_j

Name of strain j

replicate

Experimental block

input_freq_i

input frequency of strain i (determined by experiment)

input_cells_total

total number of cells plated for development (determined by experiment)

output_freq_i

output frequency of strain i (measured using flow cytometry or microscopy)

spores_total

total number of spores produced (measured using hemocytometer)

References

Madgwick PG, Stewart B, Belcher LJ, Thompson CRL, and Wolf JB (2018) Strategic investment explains patterns of cooperation and cheating in a microbe. Proceedings of the National Academy of Science USA 115: E4823-E4832. https://doi.org/10.1073/pnas.1716087115